% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_monetary_sentiment.R
\docType{data}
\name{dictionary_monetary_sentiment}
\alias{dictionary_monetary_sentiment}
\title{Monetary Sentiment Word Lists}
\format{
An object of class \code{dictionary2} of length 15.
}
\source{
\url{https://github.com/poltextlab/central_bank_communication}
}
\usage{
dictionary_monetary_sentiment
}
\description{
A \pkg{quanteda} \link[quanteda]{dictionary} object containing the 2021 version of the Monetary Sentiment Word Lists.
}
\references{
Máté Á, Sebők M, Barczikay T (2021). \href{https://doi.org/10.1371/journal.pone.0245515}{The effect of central bank communication on sovereign bond yields: The case of Hungary}.
\emph{PLoS ONE}, 16(2): e0245515.
}
\keyword{datasets}
